[DC-1756] Move util function to separate file [ci skip]
