+++ data_steward/analytics/cdr_ops/ehr_snapshot_qc.py
@@ -22,7 +22,8 @@ EHR_SNAPSHOT_DATASET_ID = ""  # Identifies the snapshot dataset
 # +
 import pandas as pd
 
-from utils.bq import get_client, execute
+from utils.bq import get_client
+from analytics.cdr_ops.utils import execute
 
 client = get_client(PROJECT_ID)
