+++ data_steward/analytics/cdr_ops/utils.py
@@ -1,0 +1,15 @@
+"""
+Utility functions for notebooks related to CDR generation
+"""
+
+
+def execute(client, query):
+    """
+    Execute a bigquery command and return the results in a dataframe
+
+    :param client: an instantiated bigquery client object
+    :param query: the query to execute
+    :return pandas dataframe object
+    """
+    print(query)
+    return client.query(query).to_dataframe()
